requirements.txt
pandas==2.0.2
pymongo==4.4.0
python-decouple==3.8
Requests==2.31.0
sendgrid==6.10.0
